refactor(vsock): split connection manager
